LayoutTests/imported/w3c/web-platform-tests/dom/events/scrolling/wheel-event-transactions-target-display-change_include=contents-expected.txt


FAIL Modify the initial wheel event target to display:contents assert_equals: Initial wheel event is has the modified element as the target expected Element node <div id="initialTarget" class="spacer" style="background:... but got Element node <div id="firstRootSpacer" class="spacer" style="backgroun...
